[bitnami/airflow] Fix the wrong env var name for Airflow fernet key
There's an issue with the fernet key deployed with the current config. The vars which
are created throught the Airflow UI cannot be properly read by all workers because
they are using different fernet keys. The reason for this is the fact that the fernet
key env var name is not the right one. It is named AIRFLOW_FERNET_KEY and it must be
AIRFLOW__CORE__FERNET_KEY according to the Airflow documentation -
https://airflow.readthedocs.io/en/stable/howto/secure-connections.html