[bitnami/airflow] Allow existingSecret for redis/postgresql .enabled = true (#2279)
* Allow  ExistingSecrets for postgres/redis enabled

For postgresql.enabled= true, and redis.enabled = true, allow the option of providing existing Secret name. 
Right now, if postgresql.enabled = true, it doesn't use existingSecret provided in externalDatabase.existingSecret & does not create a pod for postgres as well. Same logic for redis

* Add redis.existingSecret condition

generate this secret if redis.existingSecret is not provided on the top of existing conditions

* Add postgresql.existingSecret condition

* add redis.existingSecret/postgresql.existingSecret

* add redis.enabled/postgresql.enabled conditions

* Add redis.existingSecret/ posgresql.existingSecret

* add redis.existingSecret/postgresql.existingSecret

* upgrade to 5.0.7

* fix alignments

* upgrade to 5.1.0

* [bitnami/airflow] Update components versions

Signed-off-by: Bitnami Containers <containers@bitnami.com>

Co-authored-by: Marcos Bjoerkelund <marcos@bitnami.com>
Co-authored-by: Bitnami Containers <containers@bitnami.com>