[bitnami/postgresql-ha] Default pgpool loadbalancing to false
pgpool LoadBalancing is an advanced feature, and should be
disabled by default. In the current default state of the chart,
pgpool load balancing can lead to invalid read after write
synchronization errors. 

Example: 
 1. application 1 INSERTs a new row in a table. pgpool sends this to the master node
 2. application 1 advertises this new row in a message passing system, say rabbitMQ
 3. application 2 receives the message, and tries to SELECT the row. pgpool with load balancing pushes the select to one of the slaves. BUT, the slave hasn't gotten the replicated data yet. The SELECT fails with no rows.

While there are solutions to the above problem, I think it illustrates that pgpool load balancing has enough footguns
that it should be disabled by default and opt in.