[bitnami/postgresql-ha] Add postgresqlPassword option
When `postgresql.username` is not `postgres`, a new user is created that
only has permissions for the database created with `postgresql.database`
and the `postgres` user does not have a password and cannot be accessed
remotely.

This commit adds the `postgresql.postgresPassword` property that can be
used to specify the password for the `postgres` user when
`postgresql.username` is set to a different user.

Resolves #2470.