[bitnami/thanos]: Move querier Ingress config to querier section
Thanos querier is not the only component which can benefit from an ingress
object. bucketweb, storegateway, querier etc can all use an ingress
resource to be exposed outside of cluster. As such, it's cleaner to have
the ingress configuration for querier in the relevant section instead of
having it as a global configuration option. This enables us to form a
pattern on how to introduce ingresses to the rest of the resources.