[bitnami/thanos] Add Ingress for Querier GRPC service
The querier uses a GRPC service which is useful when you have stacked
Thanos deployments. For example, it's possible to have a single observer
cluster which monitors many different clusters. Each one of these
clusters, including the observer, may have its own Thanos stack running.
However, from the observer cluster, the querier needs to connect to the
each cluster's querier in order to have a global view from all the
clusters. As such, in order to do that, we need to expose the Querier
GRPC service through the ingress in order to allow cluster-to-cluster
communication.