[bitnami/redmine] Inject Root CA into image (#2774)
* [bitnami/redmine] Inject Root CA into image

Add a certificate authority into the redmine image. Required for ldaps
authentication.

Signed-off-by: Joseph Ball <joseph.ball@packetsolutions.io>

* [bitnami/redmine] Add custom certificate into pod

Insert a custom certificate into the pod, allowing for a trusted
connection using a loadbalancer, or between the ingress and the pod.

Signed-off-by: Joseph Ball <joseph.ball@packetsolutions.io>

* [bitnami/redmine] Update certificate image to use bitnami schema

Change image from alpine to minideb. Init container will test to see if
apk package manager is available for user to change source image.

Signed-off-by: Joseph Ball <joseph.ball@packetsolutions.io>

* [bitnami/redmine] Update components versions

Signed-off-by: Bitnami Containers <containers@bitnami.com>

Co-authored-by: Bitnami Containers <containers@bitnami.com>