Print offending value in validateValues
Hi,

I'm debugging why I'm hitting this issue and is really difficult because
it's a remote CI, and this chart is a dependency of my dependency,
and ... I don't want to waste your time.

My point is - if there is an invalid value that doesn't contain secrets, please print it.
Helm is such a terrible combination of templates and yamls files, with absolutely no
resistance for mistakes like typos, that in a larger setups is really difficult to track
the root cause of problems, and a printout like this at very least confirms if the
value is what we think it should be.