[bitnami/rabbitmq] Corrected example of extraContainerPorts (#3076)
* Corrected example of extraContainerPorts

The example given in #2023 and merged in #2027 specified the property `ports` which is not valid when rendered in the StatefulSet. Instead `containerPorts` should be used.

* Corrected example of extraContainerPorts

The example given in bitnami#2023 and merged in bitnami#2027 specified the property `ports` which is not valid when rendered in the StatefulSet. Instead `containerPorts` should be used.

* Corrected example of extraContainerPorts

The example given in bitnami#2023 and merged in bitnami#2027 specified the property `ports` which is not valid when rendered in the StatefulSet. Instead `containerPorts` should be used.