[bitnami/prometheus-operator] Expose additional ports from Thanos sidecar
The sidecar is configured to expose the following two ports

- GRPC/10901
- HTTP/10902

However, the service that targets the sidecar does not expose the later
making it impossible to get any metrics from it through traditional
prometheus scraping as these are only available to the HTTP endpoint.
As such, add a new option to allow us to expose additional ports from
the sidecar container.