[bitnami/prometheus-operator] Use the operator template for container name
Using the top level 'prometheus-operator.name' template can have weird
consequences in some cases. For example, imagine we use the default
value of 'prometheus-operator' as Chart name. In this case, the template
will strip the '-operator' part and the container will end up being
named 'prometheus'. However, the name 'prometheus' is also used by the
prometheus CRD when it creates prometheus instances. As such, it makes
it harder to distinguish between the prometheus-operator container and
those in the actual prometheus instances. It's better to have a clear
identification for the container that belongs to the opearator