[bitnami/prometheus-operator] Revert "Adjust thanos-sidecar for prom-op no longer using mount subpaths (#3122)"
This reverts commit ad72b27091f2b7ee29e22c236463c6a8a1b4e6b7.

The change is breaking Thanos due to disableMountSubPath option not
existing to the Prometheus CRD

Error: unable to build kubernetes objects from release manifest: error validating "": error validating data:
ValidationError(Prometheus.spec.storage): unknown field "disableMountSubPath" in com.coreos.monitoring.v1.Prometheus.spec.storage

Which is actually true since the CRD installed by Bitnami does not have
this option in the spec

https://github.com/bitnami/charts/blob/master/bitnami/prometheus-operator/crds/crd-prometheus.yaml

As such there is no way to set this option to the values.yaml file

However, the message in the original commit is rather confusing as the
option hasn't changed its default value upstream

https://github.com/coreos/prometheus-operator/blob/release-0.40/pkg/prometheus/statefulset.go#L947

The operator will still return 'prometheus-db' subpath by default. Also
this option was not documented at all in the README file so it's not
clear what its intention is.

Having said that, not being able to override it in the values.yaml file,
means that Thanos sidecar is unable to mount the subpath no matter what
and there is no straight way to recover this apart from editing the live
objects.

As such, I suggest we revert this in order to restore functionality as
it was until it's clear what problem we are trying to solve.