Variant2: Implement certgen by using hooks (#2)
* Implement certgen via hooks
* Delete resources with helm hooks only if needed
* Revert "Delete resources with helm hooks only if needed"

This reverts commit 76449252d06b7bda0f16c490316478c3fb1004f1.

As documented inside the helm docs, we should remove unneeded resources:
~~~
The resources that a hook creates are currently not tracked or managed as part
of the release. Once Helm verifies that the hook has reached its ready state,
it will leave the hook resource alone. Garbage collection of hook resources
when the corresponding release is deleted may be added to Helm 3 in the future,
so any hook resources that must never be deleted should be annotated with
`helm.sh/resource-policy: keep`.

Practically speaking, this means that if you create resources in a hook, you
cannot rely upon helm uninstall to remove the resources. To destroy such
resources, you need to either add a custom `helm.sh/hook-delete-policy`
annotation to the hook template file, or set the time to live (TTL) field of
a Job resource.
~~~