[bitnami/redis] Add exec to forward signals to redis
Calling the next script without exec is causing a fork, which means all
the signals that are being sent to the Pod are lost.

This is especially dangerous for databases such as Redis, because
handling the SIGTERM when Kubernetes needs to scale down the pod is
critical for ensuring a clean shutdown and data integrity.

Adding `exec` ensures that we're not forking, and the signals work as
intended.

More reading:
- https://hynek.me/articles/docker-signals/
- https://ss64.com/bash/exec.html