[bitnami/*] Extra manifests should be top-level
extra manifests declared with the `extraDeploy` value used to be defined in a single List manifest.
the issue with this setup, is that if you want to define a Job to be used as a Helm hook, it won't work - because Helm only look at top-level manifests to find the hook annotations.

this PR changes the implementation of extra manifests, by rendering top-level manifests instead of wrapping them inside a List.
charts users don't need to change anything in their values.yaml