[bitnami/dokuwiki] Fix multiple ingress hosts
Previously, when attempting to use multiple ingress hosts, only the last ingress was created.
This was because the template was appending a resource template with the same name,
causing only the last one to be enabled.

Iteration has been moved inside the resource template to fix this issue and support
multiple ingress hosts.

Fixes #4618