Allow release to specify namespace for install
The deployment, service, and serviceaccount resources were missing
metadata.namespace which limited options for installing the chart to a
specific namespace. Now the namespace is specified for all resources and
using the --namespace command with helm install isn't required anymore,
which is helpful when using tools like Argocd or Grafana Tanka.