Update instruction for nodes in values.yaml
It might be a case, where people will override the node count and will set to something less than 3 and will not necessarily understand the inclusion of replica count. More importantly, if the node count does not allow at least 3 master nodes to be created e.g. nodes=3/4 with replicas=1, the cluster creation will fail.

This has to be clearly mentioned to avoid further confusion.