kibana savedObjects not processed when server.rewriteBasePath
Address issue #6455 by including the basePath value in the API URL which is used to push saved objects into kibana.

Additionally, add support for ingress.hostname: "*" so that an existing ingress controller can be used to make kibana available.   The current support for ingress.hostname requires a known hostname to be matched by the ingress service.  If the ingress service only has an IP address, then there is now way for the kibnana ingress to be accessed.   The hostname "*" is not a valid specification, so using it to suppress the generation of the hostname field on the ingress, solves the problem without introducing any incompatibilities with the previous default hostname kibana.local.