[bitnami/grafana] Allow use of ingress without a hostname
Allow use of an ingress without a known hostname. It allows the hostname rule to be dropped by use of the "*" value in the grafana chart consistent with a recent change in the kibana chart.  The value "*" is not valid for a hostname input, so it does not introduce any incompatibly with prior behavior, or ambiguity.