[bitnami/redis]: Add an extra VolumeMount to the metrics sidecar for custom scripts
If `metrics.extraArgs.script` is defined, add an extra VolumeMount
to the metrics sidecar containers to ensure they can see it.  The
responsibility of providing the volume and script is on the chart,
by way of the new `metrics.extraVolumes` and `metrics.extraVolumeMounts`
parameters.

This was not added to the sentinel sidecar, as that is apparently
going away.

Fixes issue: https://github.com/bitnami/charts/issues/6792