kube-state-metrics.serviceMonitor.honorLabels true
Setting the default value to true for honorLabels to ensure that the
pods scraped by kube-state-metrics have their labels as the ones set
inside prometheus rather than the labels on the kube-state-metrics pod.

For example, without this setting, container=kube-state-metrics
exported_container=YOUR_APP rather than container=YOUR_APP which is
confusing when triggering alerts off of the back of this.