[bitnami/postgres-ha] Do not enforce init containers if volumePermissions is disabled
There is no need to use a init container with a special security context
for chmod's if a extendedConf* is used and volumePermissions are
disabled. If the init container is forced the deployment of the
statefulset will fail on e.g. openshift with a error message like below.

...pods "postgresql-ha-postgresql-1" is forbidden: unable to validate
against any security context constraint:
[spec.initContainers[0].securityContext.runAsUser: Invalid value: 0:
must be in the ranges:...