Set namespace for all resources in rabbitmq-cluster-operator chart (#7911)
Currently the templates of ServiceAccount and the ServiceMonitor do not render
the namespace. With `helm install` this is not an issue because helm will
apply those resources to the namespace that is passed to it.
But if you use `helm template` it does not work. By adding the namespace to the
metadata of all resources the chart works with `helm install` and
`helm template` equally.

Signed-off-by: Max Rosin <max.rosin@hetzner-cloud.de>