[bitnami/redis-cluster] Add namespace in metadata (#8821)
The scope of this work is to add namesapce into all the templates used
to generate the yaml files required to deploy all the components of this
chart, if enabled.

This pretty much replicate the work from the redis chart, with the
adjustments required and all the parts dedicated to this chart
completed.

We migrated from the redis chart (using sentinel) and we noticed that
this one was missing the namespace in the metadata object.
We added it, tested it and it is running in our dev environment for a
while, so we believe that this is correctly done.

Also the tests and the template generated looks ok.

We have generated a template with a "test-ns" namespace and all the
components enabled, not completely filled, except for the ones that were
affected by the scope of this work.

We hope you like this work.

The engineering YData team <developers@ydata.ai>

Signed-off-by: Luís Portela Afonso <luis.portela@ydata.ai>
Signed-off-by: bbaudelet <bbaudelet@gmail.com>