add a service for each thanos-receive pod.
When using thanos-receive in dual mode, it is currently hard to define
a hashring, as the IPs of the pods are not known before deployment.
Adding a service for each pod, allows to address the pods via their name
in the hashring config, allowing a deployment that works without later
adjustments to the hashring config.

Signed-off-by: Arne Claus <arne.claus@trivago.com>