remove replicas setting if autoscaling is enabled.
When an HPA is present, it will override the replicas setting based on the status
of the HPA. When a deployment or statefulset is deployed again, it will override
the current HPA value with the value set in the deployment. If that value is 1
and the HPA is currently at, let's say 10, this can cause majour issues.

Signed-off-by: Arne Claus <arne.claus@trivago.com>