[bitnami/common] Remove quote in common.secrets.passwords.manage
This commit removes the `quote` function from the definition of the
`password` variable in `common.secrets.passwords.manage`.

With `quote`, I'm constantly getting a diff (in my deployment of
bitnami/harbor version 12.3.1) because the applied secret value never
has quotes while the template do have quotes, so `helm diff` reports a
diff despite there's nothing to apply really.

Since the value is base64, it should be safe to not quote it.

Signed-off-by: Andreas Lindhé <andreas@lindhe.io>