[bitnami/harbor] Add parameters to configure tracing
Tracing support was added to Harbor in version 2.4.0.

Currently the way to enable this with the Helm chart is to set
extraEnvVars for the core, jobservice and registry controller.

This gives us an easier way of enabling tracing, without having to
manually figure out which environment variables need to be added.

I've tried to keep code duplication to a minimum by templating the
environment variables and including them in the envvar configmaps for
each of the deployments. Since the registry controller did not have
a envvars configmap I added one.

Signed-off-by: Kristinn Björgvin Árdal <kristinnardalsecondary@gmail.com>