[bitnami/mediawiki] Don't generate certificates if cert-manager annotation is found
ingress.certManager is deprecated, yet required to use cert-manager.
When it's used it prevents creation of a self signed certificate and
places an optional annotation kubernetes.io/tls-acme: "true" on the
ingress.

This change auto-detects (using the bitnami common chart) the
cert-manager annotation(s) to prevent creation of the self-signed cert
and helps users follow the deprecation guidelines of using their own
annotations instead of relying on ingress.certManager.

The deprecated behavior is retained if ingress.certManager is
specified.

Signed-off-by: Eric Stokes <fernferret@gmail.com>