[bitnami/mongodb-sharded]Fixed a few errors in the templates that support monitoring, PDB, etc (#13482)
* Added missing space
Without this space the template function to create a common serviceaccount errors out.

* Added missing '$' for referencing Values
When enabling the PDB saw errors adding Labels.

* Added missing label selector for PodMonitors
This template adds PodMonitors for each shard but was broken when deploying more than 1 shard, it was missing the delimiter "---" and the label selector would have found all shards, by adding the extra label to the selector it works as expected.

* Updated Chart version

Signed-off-by: cten <cxtenberge5@gmail.com>
Signed-off-by: Kazimieras <kazimieras@omnisend.com>