zookeeper: Don't re-initialize keystores after success
Init containers must be idempotent because they may get re-run in some
cases even if the pod itself is not being restarted. For example, if
the underlying node garbage collects the associated docker container.

In the current incarnation of the init script we will fail if the init
container is re-run resulting in a state where the init container is
continually flapping.  To avoid this, we can add a initialization
check file and short circuit the initialization process if it is
present.

Signed-off-by: Michael Kelly <mkelly@arista.com>