[bitnami/redis] Sentinel deployment: Fix Failover at graceful shutdown (#14133)
* [bitnami/redis] Sentinel deployment: Fix Failover at graceful shutdown

Achieving a proper failover at graceful shutdown of the redis master
pod is tricky (see #13021 and #12598 for discussions of different
options).

The currently implemented solution does not work as intended because
of a syntax error (see #13956). After fixing the syntax error, it
turns out that the intended solution does not work, as it issues a
"SENTINEL FAILOVER" as well as a Redis "FAILOVER". These failover
commands are unrelated and were not designed to work
together. Problems are for example:

1. Redis "FAILOVER" and "SENTINEL FAILOVER" may select different new
   master nodes.

2. As part of Redis "FAILOVER", the master will wait for lagging
   replicas. But while waiting, the replicas may "disappear" because
   the sentinel failover procedure is reconfiguring them. This causes
   the redis failover procedure to hang until it times out.

In summary, the Redis "FAILOVER" command can't be used in this
context.

Instead, implement the following when the redis master pod shuts down:

1. Issue a "CLIENT PAUSE" command on the Redis master. This will block
   all writes. Reads can still happen.

2. Trigger a failover to a different master by issuing a "SENTINEL
   FAILOVER" command.

3. The shutdown of the sentinel container is delayed until the
   sentinel returns a different master (i.e. the failover has
   succeeded from a client's perspective)

   Optionally (the default), also delay the shutdown of the redis
   container until that point.

The delayed shutdown of the Redis container maximizes the availability
of Redis during the failover (at least for reads). However, as
discussed in
https://github.com/bitnami/charts/pull/13021#issuecomment-1289177976,
it has drawbacks when short timeouts are used by clients and can
therefore be switched off by setting
`sentinel.redisShutdownWaitFailover=false`.

Note: Because of the syntax error in the previous code, its behavior
actually corresponds to the behavior implemented by this commit when
`sentinel.redisShutdownWaitFailover=False`.

Closes #13956

Signed-off-by: Simon Baatz <gmbnomis@gmail.com>

Signed-off-by: Simon Baatz <gmbnomis@gmail.com>
Signed-off-by: Fran Mulero <fmulero@vmware.com>
Co-authored-by: Fran Mulero <fmulero@vmware.com>
Signed-off-by: Eduardo Morillo <eduardo.morillo@blueamber.io>