Add forceS3Https to values.yaml
This adds a boolean namedforceS3Https to values.yaml.  This forces Mastodon's
S3_PROTOCOL to always be set to https if true.  The default value is false.

This was added for when minio is used instead of external S3 storage.
There already exists minio.tls.enabled which requires minio.auth.existingSecret
or tls.autoGenerated to be set.  The issue here is that minio.auth.existingSecret
sets mastodon.s3.secretName which also requires Values.externalS3.existingSecretAccessKeyIDKey
be set or most deployments break.  If minio.tls.autoGenerated is used the actual cert you may
have won't be used.  This option accounts for an exisiting secret correctly usable in front
of container and have mastodon-web correctly generate https links for S3 stored content if
minio is used.

In the usecase where cert-manager is already
being used to provision and manage TLS certificates and/or nginx-ingress is
used you might not want to use another ingress which may cost another
LoadBalancer IP.

forceS3Https should only be use when TLS is needed and you already have a valid
TLS certificate that a browser will be able to validate. If S3_PROTOCOL is set to
http then mastodon-web generates http imag links that violate Content Security Policy
in a user browser.

Signed-off-by: Joel Parker <joelsparker@gmail.com>