Add forceHttpsS3Protocol to values.yaml with a default value of false
This setting will allow S3_PROTOCL in mastdon's default configmap to be
set to https independent of minio.tls.enabled's setting.

The reason?  You may already have a managed/provisoned TLS cert with
and ingress terminating TLS (cert-manager with nginx-ingress for example)
where you want minio exposed as only a ClusterIP service and not TLS at the
container.  For this to happen and for mastodon to create links to S3 stored
data (profile images, media) S3_PROTOCOL needs to be https and minio as a Pod
needs to be http only.

Signed-off-by: Joel Parker <joelsparker@gmail.com>