[#14600] Keep intermediate certs in PEM server certificate chain
The setup.sh script removed all intermediary certificates
from the configured server certificate chain provided in a tls.crt
PEM file as created by k8s cert-manager.

This effectively prevented the kafka broker to include any of the
intermediary certificates in a TLS handshake with a peer/client.

The PEM file is now being copied as-is to the kafka container.
However, this change alone does not fix the issue yet because the
kafka_configure_ssl() function from the "libkafka.sh" bitnami kafka
container helper library transforms the keystore PEM file to a
multi-line string to be include in the server.properties in a way
that is not compatible with the regular expression used by kafka's
DefaultSslEngineFactory$PemParser.pemEntries() method when starting
up the broker. The regular expression only matches the last
certificate in the chain which leads to the broker only serving
this certificate in TLS handshakes.

Signed-off-by: Kai Hudalla <sophokles.kh@gmail.com>