[bitnami/keycloak] Use `httpRelativePath` for ingress path and service monitor endpoints
When `httpRelativePath` is set, keycloak listens on that path. If a user sets `httpRelativePath` they (optionally) need to match the same in `ingress.path` and must (if they use metrics) match the same path on `metrics.serviceMonitor.endpoints[]`.

This change automates setting these manually when setting `httpRelativePath` and improves user experience.

Signed-off-by: Yarden Shoham <git@yardenshoham.com>