[bitnami/apache] Remove ingress.certManager
This change removes the long deprecated annotation for cert-manager from
the Ingress resource.

NOTE: Technically, this is a breaking change. If someone is using the
`ingress.certManager` value, they will lose the annotation. I imagine
there are few cases where the annotation has any effect, but there's
always someone out there running some really odd configuration...

It's of course still possible to have the annotation, it just has to be
included in `ingress.annotations`, along with the rest of the
annotations.

Since the value is no longer in use anywhere, we also remove it from
values.yaml.

Signed-off-by: Andreas Lindhé <andreas@lindhe.io>