[bitnami/opensearch] Fix node-role guards checking autoscaling
There are guards around Kubernetes resources specific to nodes
differentiated by OpenSearch roles. These guards check whether
particular role is enabled. We have implemented these guards, so they
check at least if autoscaling is enabled or if replica count is greater
than zero. However, we support both horizontal pod autoscaler (HPA) and
vertical pod autoscaler (VPA). Our Helm chart values have no unspecified
autoscaling.

We change guards to check if HPA is enabled. If it is enabled (and other
conditions are true), we should configure Kubernetes resources
associated with particular role. This is because pods for this
particular role will be created as needed.

We do not care about VPA, because VPA requires some pods to exists in
the first place and we already cover that (or fix in this commit).

Signed-off-by: Filip Zrůst <filip.zrust@sinsir.net>