[bitnami/opensearch] Fix guards for OpenSearch Dashboards (#18393)
* [bitnami/opensearch] Fix node-role guards checking autoscaling

There are guards around Kubernetes resources specific to nodes
differentiated by OpenSearch roles. These guards check whether
particular role is enabled. We have implemented these guards, so they
check at least if autoscaling is enabled or if replica count is greater
than zero. However, we support both horizontal pod autoscaler (HPA) and
vertical pod autoscaler (VPA). Our Helm chart values have no unspecified
autoscaling.

We change guards to check if HPA is enabled. If it is enabled (and other
conditions are true), we should configure Kubernetes resources
associated with particular role. This is because pods for this
particular role will be created as needed.

We do not care about VPA, because VPA requires some pods to exists in
the first place and we already cover that (or fix in this commit).

Signed-off-by: Filip Zrůst <filip.zrust@sinsir.net>

* [bitnami/opensearch] Fix guards for OpenSearch Dashboards

We already declare Helm chart value to enable or disable OpenSearch
Dashboards. We also use guards checking similar and other values for
other types of workloads. We use these guards to ultimately decide if
particular workloads and Kubernetes resources associated with them
should be configured. What we have been missing is a similar guard which
checks if OpenSearch Dashboards should be configured.

We fix this by introducing opensearch.dashboards.enabled helper template
definition. It is similar to opensearch.ingest.enabled because Helm
chart values configuring ingest contain “enabled” field as well as
fields all the guards check: autoscaling.hpa.enabled and replicaCount.
We implement equivalent condition for the new helper template
definition.

We also apply this new guard consistently with other types of workloads…

Signed-off-by: Filip Zrůst <filip.zrust@sinsir.net>

* [bitnami/opensearch] Bump version after OpenSearch Dashboards fix

Signed-off-by: Filip Zrůst <filip.zrust@sinsir.net>

---------

Signed-off-by: Filip Zrůst <filip.zrust@sinsir.net>
Signed-off-by: Ibone González Mauraza <gibone@vmware.com>
Co-authored-by: Ibone González Mauraza <gibone@vmware.com>