[bitnami/common] new macro to checksum resources
The pattern to rollout workloads on config change, described in
[helm docs](https://helm.sh/docs/howto/charts_tips_and_tricks/#automatically-roll-deployments),
has the issue that it will trigger a reload of the pods when the chart version
change, even it the pod template and configmap data are not changed.

This happens because the version of the chart is included in "helm.sh/chart"
in metadata.labels (printf "%s-%s" .Chart.Name .Chart.Version) of the configmap
and the checksum is done on the whole file.

This new macro calculates the checksum by omitting metadata.

It works only if there is a single resource in the file (limitation of
fromYaml)

Signed-off-by: Cyril Jouve <jv.cyril@gmail.com>