[bitnami/zookeeper] Make Zookeeper DefaultMode YAML 1.2 Compliant (#21081)
* Make Zookeeper DefaultMode YAML 1.2 Compliant

With the YAML 1.2 spec octals must be prefixed with `0o` and not `0`.
Parsers implementing the latest spec run into issues. See discussion
here: https://github.com/dtolnay/serde-yaml/pull/225

This commit makes the Zookeeper script mount compliant with YAML 1.2.

Signed-off-by: Marc Schreiber <info@schrieveslaach.de>

* Use Decimal Notation for defaultMode

In order to be still YAML 1.1 compliant and to be YAML 1.2 compliant,
this commit changes the defaultMode to the decimal value 493. This seems
to be done in other places as well (see modes of TLS secrets).

Signed-off-by: Marc Schreiber <info@schrieveslaach.de>

---------

Signed-off-by: Marc Schreiber <info@schrieveslaach.de>