[bitnami/kafka] embedded jmx exporter agent
Run the jmx exporter for prometheus as an embedded agent in the Kafka
broker JVM, rather than as a standalone process.

As explained in https://github.com/prometheus/jmx_exporter, this is
the preferred approach. The most obvious benefit is that it then also
exposes JVM process metrics (e.g. about memory usage etc).

Signed-off-by: Jan Van Besien <janvb@ngdata.com>