[bitnami/rabbitmq-cluster-operator] feat: 🔒 support for extra rules
The existing .clusterOperator.rbac.customRules is an override to RBAC
permissions in the Cluster Operator ClusterRole. Sometimes, a user
simply needs to append some rules, specific to their environment. For
example, to deploy in an environment with Pod Security Policies enabled;
in such environment, the ClusterRole needs an extra rule to allow usage
of PSPs.

The current implementation is "one or the other", meaning that
customerRules take precedence over extraRules.

[#22248]

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>