Fixed jupyterhub external database config gen
This chart had an issue where the configuration for external databases
was both behind a flag enbabling the internally bundled db, and what
appeared to be copy and pastes gone wrong where variables were pointing
to entirely inappropriate values. I also noted that the affected lines of
template had not been touched in years.

This was tested with bitnami/postgresql-ha under an umbrella chart to
ensure both the external database and jupyterhub were deployed under
minikube.

Changelog: Bugfix
Signed-off-by: George Raven <GeorgeRavenCommunity@pm.me>