[bitnami/discourse] Update plugins to compatible versions
This change introduces a new helm value `discourse.compatiblePlugins` which is per default enabled.

If enabled and the plugins are persisted (`discourse.persistPlugins`) it will update all plugins to the latest version (same as on first time installation) on initialization.

Then it will update all plugins to a compatible version (for upgrade and first time installation).

This change should fix upgrades with outdated plugins and fix install if a plugin already has a newer version.

Signed-off-by: Gernot Pansy <gernot@pansy.at>