[bitnami/redis] Fix "resources" sections warning for sentinel deployment
For a sentinel deployment without configured "resources", the notes of the helm release contain
the following warning:

```
WARNING: There are "resources" sections in the chart not set. Using "resourcesPreset" is not recommended for production. For production installations, please set the following values according to your workload needs:
  - master.resources
  - replica.resources
  - sentinel.resources
+info https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
```

This is misleading as `master.resources` isn't used at all in a sentinel deployment.

Fix this by adding "master" to the list of checked sections conditionally.

Signed-off-by: Simon Baatz <gmbnomis@gmail.com>