Remove Skipper ConfigMap and Secrets when Skipper is disabled
This change was made to fix a discovered issue from the values
documentation mentioning that `externalDatabase.existingPasswordSecret`
and `externalDatabase.existingPasswordKey` is deprecated in favor of
using `existingSecret` and `existingSecretPasswordKey` within `dataflow`
and `skipper`. When changing to the expected way but having skipper
disabled, the skipper `externaldb-secret.yaml` is still included and
fails to base64 encode the password since it is empty. The best solution
was not to include the file at all. I applied the same logic to skippers
`configmap.yaml` since it is not used when enabled.

Signed-off-by: Anthony Lofton <anthony.lofton@jbhunt.com>