[bitnami/keycloak]: add url env vars to fix admin ingress access
Replace KC_HOSTNAME_ADMIN by KC_HOSTNAME_ADMIN_URL to include the scheme
(http or https) and add KC_HOSTNAME_URL for the frontend to function
correctly. These variable allow the front to function correctly if we
have the tls option enabled. Before this commit the front was always
attempting to load some resources over http while it may be exposed in
https.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@ledger.fr>