rabbitmq-cluster-operator: rename http-webhook port to https-webhook
It's actually https, not http. Istio does magic based on the name of the port and
expects http traffic on http-<name> ports, not https.
This results in a "connection reset by peer" when trying to deploy a
RabbitmqCluster resource, when it tries to invoke the message topology
operator's webhook.

Signed-off-by: Boris Brönner <broenner@xelonic.com>