postgres exporter monitor all databases
Currently there is no option to monitor all databases with postgres exporter without mentioning the databases names and creating them in the init scripts. (.Values.auth.database). Postgres Exporter does let this feature work if you dont mention any databases in the URI and that is why i think the required statement is not correct so i removed it.

Signed-off-by: Koren Peretz <62677694+KorenP1@users.noreply.github.com>