postgres exporter monitor all databases (#27586)
* postgres exporter monitor all databases

Currently there is no option to monitor all databases with postgres exporter without mentioning the databases names and creating them in the init scripts. (.Values.auth.database). Postgres Exporter does let this feature work if you dont mention any databases in the URI and that is why i think the required statement is not correct so i removed it.

Signed-off-by: Koren Peretz <62677694+KorenP1@users.noreply.github.com>

* Update Chart.yaml

Adding Chart.yaml version bump to pass tests

Signed-off-by: Koren Peretz <62677694+KorenP1@users.noreply.github.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

* Update statefulset.yaml

Looks like there is no usage of the DB NAME part in the DATA_SOURCE_URI environment variable which means there is no need for the $database part, the postgres exporter always monitoring all the dbs, also if created after the initialization and not mentioned in the DATA_SOURCE_URI env var.

In conclusion the PR is for two reasons.
1: making DATA_SOURCE_URI env var more "right" and related to what it actually does
2: More importantly, removing the required statement in the $database helm variable assignment in order for users to be able to deploy the postgres exporter although not assigning .Values.auth.database variable for the db creation in the initialization of the postgres

Signed-off-by: Koren Peretz <62677694+KorenP1@users.noreply.github.com>

* Update statefulset.yaml

Signed-off-by: Koren Peretz <62677694+KorenP1@users.noreply.github.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

---------

Signed-off-by: Koren Peretz <62677694+KorenP1@users.noreply.github.com>
Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>
Co-authored-by: Bitnami Containers <bitnami-bot@vmware.com>