[bitnami/harbor] initContainer for certificate vol
Add an init container to copy the /etc/ssl/certs directory to the
ephemeral volume using the same user and group that the main container
runs as.
Then mount that volume over the /etc/ssl/certs directory within the main
container so that the /etc/ssl/certs/ca-certificates.crt can be updated
when internalTLS.enabled is set to true.

Signed-off-by: Jim Barber <jim.barber@healthengine.com.au>