Update prometheusrules.yaml
This commit removes a conflicting label from the prometheusrules template.
Two labels with the key 'app.kubernetes.io/component' existed. This is not possible, and caused issues with FluxCD. This change removes the first label 'app.kubernetes.io/component=nginx' as it was overwritten with 'app.kubernetes.io/component=metrics' during manual installs.

Signed-off-by: Hans Strijker <45356184+HStry@users.noreply.github.com>