[bitnami/nginx] Release 18.1.12 - bugfix prometheusrules label collision
This commit removes a conflicting label from the prometheusrules template.
Two labels with the key 'app.kubernetes.io/component' existed. This is
not possible, and caused issues with FluxCD. This change removes the
first label 'app.kubernetes.io/component=nginx' as it was overwritten
with 'app.kubernetes.io/component=metrics' during manual installs.

The version has likewise been updated to reflect the change.

Signed-off-by: Hans Strijker <45356184+HStry@users.noreply.github.com>