Contour: Expose NodePort for the metrics port
When deploying Contour behind an existing load balancer, such as an AWS ALB, health checks are often required for the target group.

Currently, the Envoy health check endpoint is not exposed, making it unavailable for such checks. This change adds the option to expose the metrics port as a NodePort, enabling health checks to be performed.

Signed-off-by: Mario de Frutos <mario@defrutos.org>