adds metrics port to values and uses in statefulset, metrics service and network policy
Starting with Keycloak 25 the metrics port was switched from 8080/8443 to 9000. This commit introduces a metrics port in the containerPorts configuration. This metrics port is used to allow ingress via the network policy, expose the container port and also as target for the metrics service. The port of the metrics service itself is not changed. Therefore, metrics are still exposed on port 8080 to the outside. If Keycloak is started with KC_LEGACY_OBSERVABILITY_INTERFACE set, the containerPorts.metrics value can be set to 8080 restoring the old behaviour completely.

Signed-off-by: Martin Schmidt <maschmi@maschmi.net>