[bitnami/kibana] Fix server.publicBaseUrl handling
Previously, if .Values.extraConfiguration."server.publicBaseUrl" was
specified (with a dot in the key), that key was inserted into the
configmap twice, causing Kibana to error out.

Furthermore, since the introduction of the fallback to the ingress
hostname, this setting couldn't be templated anymore (which was possible
before).

This commit allows the templating as well as specifying it with the dot
notation.

Signed-off-by: Jan-Philipp Litza <jplitza@users.noreply.github.com>