feat(harbor)!: Upgrade postgres to version 17.2.0
The commit 017c8c368e4740be7e4d465cbb9b0c54b2ca829a should update the postgres version to version 17.x. But it did not. The patch release after this commit upgrades Postgres to version 14 and not 17. This is also referenced in the README.md of the chart. To solve this issue, this commit really updates Postgres to version 17.2.

Signed-off-by: Lukas Jost <lukas.jost@outlook.com>