Redis Sentinel - handle SIGTERM in `kubectl-shared` container
When Redis Sentinel has the master service option enabled, the `kubectl-shared`
container runs a script with an infinite `while+sleep` loop.

This script does not handle SIGTERM and causes the pod to hang upon termination
until the termination grace period elapses.

This change introduces a `preStop` hack to create a `terminate` file, which is used
as a singal to the script that it should exit the infinite loop

Signed-off-by: Kevin Pullin <kevin.pullin@gmail.com>