fix: change svc target port in servicemonitor (#32236)
* fix: change svc target port in servicemonitor

The KeyDB service port name did not match the expected name in the ServiceMonitor configuration.  
This mismatch caused Prometheus to fail to scrape metrics from the KeyDB instance.  

This commit updates the ServiceMonitor port name to ensure consistency between the Service and ServiceMonitor,  
allowing Prometheus to correctly discover and collect metrics.

Signed-off-by: Aioramu <31703899+Aioramu@users.noreply.github.com>

* increase chart version after fix

Signed-off-by: Aioramu <31703899+Aioramu@users.noreply.github.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>

---------

Signed-off-by: Aioramu <31703899+Aioramu@users.noreply.github.com>
Signed-off-by: Bitnami Containers <bitnami-bot@vmware.com>
Co-authored-by: Bitnami Containers <bitnami-bot@vmware.com>