[bitnami/keycloak] fix: preserve timestamps in init script
With the latest Keycloak version, the timestamps of providers is compared to the timestamp of the latest provider build.

When starting keycloak with the `--optimized` flag, keycloak will throw an error when the timestamp of the providers is more recent than it was during the build.

> A provider JAR was updated since the last build, please rebuild for this to be fully utilized.

As we do not preserve the timestamp in the init script, this error is always thrown when starting keycloak as optimized.

This is fixed by preserving the timestamp in the initContainer `cp` commands.

see: https://github.com/keycloak/keycloak/issues/38893

Signed-off-by: msvechla <m.svechla@gmail.com>
Signed-off-by: Marius Svechla <m.svechla@gmail.com>