[bitnami/mastodon] fix broken PDB Label matching
closes: #34544

PDB Selectors try to match for labels that are not actually present on
the Pods, but only on their Deployments, causing the Selector to quietly match nothing.

The "part-of" labels are present on almost all other objects managed
by this chart, and part of the standard Labels mentioned in Helm's best
Practices. So adding them to the pods seems like the most
consinstent solution

Signed-off-by: Stefan Neubauer <github@sneubauer.com>