feat(milvus): Add GCS support to _helpers.tpl
modified _helpers.tpl file in milvus templates to support gcs storage.
The wait-for-s3 init container expects 'RequestID' in the body of the response from curl requests for AWS S3.
However, this does not apply to GCS since it looks for 'x-guploader-uploadid' in the header response. This change
adapts the logic to properly handle GCS responses.

Signed-off-by: Abdullah Alfaifi <100507587+asalfaifi@users.noreply.github.com>