Add the option to use a password for tls PEM keys when not autogenerated.
This commit tries to keep in line with how the main container optionally allows a user to bring their own password encrypted PEM key. The password is mounted from a secret as an ENV and is passed to the openssl pkcs conversion function. If it is not present, the value will be nil and not affect the conversion process.

Signed-off-by: Abdiel Olmos <66489738+ItsCatman@users.noreply.github.com>