[bitnami/redis] Avoid eager evaluation of ternary when resolving user secrets (#36265)
* [bitnami/redis] Fix: avoid eager evaluation of ternary when resolving user secrets

Helm’s ternary function (from Sprig) is eagerly evaluated, which
caused `include "common.secrets.get"` to run even when
`.Values.auth.acl.userSecret` was empty. This resulted in errors
when the referenced Secret did not exist.

Replaced ternary with an explicit `if` guard so that secret lookup
is only performed when a userSecret is actually set.

Signed-off-by: Sina Safari <sina.devel@gmail.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Bot <bitnami.bot@broadcom.com>

* Update CHANGELOG.md

Signed-off-by: Bitnami Bot <bitnami.bot@broadcom.com>

---------

Signed-off-by: Sina Safari <sina.devel@gmail.com>
Signed-off-by: Bitnami Bot <bitnami.bot@broadcom.com>
Co-authored-by: Bitnami Bot <bitnami.bot@broadcom.com>